combined login and signup into auth route
